+++ src/app/src/main/kotlin/org/vaccineimpact/reporting_api/app_start/Routing/VersionRouteConfig.kt
@@ -41,8 +41,7 @@ object VersionRouteConfig : RouteConfig
             Endpoint("/reports/:name/versions/:version/all/", versionController, "getZippedByNameAndVersion",
                     ContentTypes.zip)
                     .allowParameterAuthentication()
-                    // more specific permission checking in the controller action
-                    .secure(),
+                    .secure(readReports),
 
             Endpoint("/reports/:name/versions/:version/publish/", reportController, "publish",
                     method = HttpMethod.post)
